♻️  core: Consolidate mock socket impls for tests
