+++ zlink-core/src/test_utils/mod.rs
@@ -1,0 +1,7 @@
+//! Testing utilities for zlink-core.
+//!
+//! This module provides mock implementations and testing utilities that are shared
+//! across different test modules in the crate.
+
+#[cfg(test)]
+pub(crate) mod mock_socket;
